pyproject.toml
[project]
name = "youtube-video-summarizer"
version = "1.0.0"
description = "A Python application that extracts audio from YouTube videos and transcribes them to text using Google Gemini API"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["youtube", "transcription", "audio-to-text", "gemini", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-genai>=1.49.0",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2025.10.22",
]

[project.urls]
Homepage = "https://github.com/yourusername/youtube-video-summarizer"
Documentation = "https://github.com/yourusername/youtube-video-summarizer#readme"
Repository = "https://github.com/yourusername/youtube-video-summarizer"
Issues = "https://github.com/yourusername/youtube-video-summarizer/issues"

[project.scripts]
youtube-summarizer = "src.youtube_summarizer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_summarizer"]
